Extend psalm over /tests and fix a few new detections
Some of the detections by psalm are cause by "imperfect"
phpdocs in the Symfony 6 upstream code. I've done my
best to fix them by adding some conditions here and there
(that allow psalm to pass because that way types are checked).

Only as last resort we'll be using PSAlm annotations to
exclude some errors. Alternative is to create baseline file
but better if we don't need any.